+++ starkbank/event/__event.py
@@ -53,7 +53,7 @@ _resource = {"class": Event, "name": "Event"}
 
 def get(id, user=None):
     """# Retrieve a specific notification Event
-    Receive a single notification Event object previously created in the Stark Bank API by passing its id
+    Receive a single notification Event object previously created in the Stark Bank API by its id
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"
     ## Parameters (optional):