+++ starkbank/transaction/__transaction.py
@@ -61,7 +61,7 @@ def create(transactions, user=None):
 
 def get(id, user=None):
     """# Retrieve a specific Transaction
-    Receive a single Transaction object previously created in the Stark Bank API by passing its id
+    Receive a single Transaction object previously created in the Stark Bank API by its id
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"
     ## Parameters (optional):