+++ starkbank/transfer/log/__log.py
@@ -32,7 +32,7 @@ _resource = {"class": Log, "name": "TransferLog"}
 
 def get(id, user=None):
     """# Retrieve a specific transfer.Log
-    Receive a single transfer.Log object previously created by the Stark Bank API by passing its id
+    Receive a single transfer.Log object previously created by the Stark Bank API by its id
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"
     ## Parameters (optional):