+++ starkbank/utilitypayment/__utilitypayment.py
@@ -56,7 +56,7 @@ def create(payments, user=None):
 
 def get(id, user=None):
     """# Retrieve a specific UtilityPayment
-    Receive a single UtilityPayment object previously created by the Stark Bank API by passing its id
+    Receive a single UtilityPayment object previously created by the Stark Bank API by its id
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"
     ## Parameters (optional):
@@ -69,7 +69,7 @@ def get(id, user=None):
 
 def pdf(id, user=None):
     """# Retrieve a specific UtilityPayment pdf file
-    Receive a single UtilityPayment pdf file generated in the Stark Bank API by passing its id.
+    Receive a single UtilityPayment pdf file generated in the Stark Bank API by its id.
     Only valid for utility payments with "success" status.
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"